Fix harvest and Action list :D
