+++ Game/Assembly-CSharp.csproj
@@ -61,7 +61,7 @@
      <Compile Include="Assets\Scripts\ObjectInfo.cs" />
      <Compile Include="Assets\Scripts\ResourceManager.cs" />
      <Compile Include="Assets\Scripts\TaskList.cs" />
-     <Compile Include="Assets\Scripts\Unit.cs" />
+     <Compile Include="Assets\Scripts\Units.cs" />
  <Reference Include="UnityEngine">
  <HintPath>G:/2019.4.11f1/Editor/Data/Managed/UnityEngine/UnityEngine.dll</HintPath>
  </Reference>
