+++ Game/Assets/Scripts/ActionList.cs
@@ -5,18 +5,18 @@ using UnityEngine.AI;
 
 public class ActionList : MonoBehaviour
 {
-    public void Move(NavMeshAgent agent, RaycastHit hit, TaskList task )
+    public void Move(NavMeshAgent agent, RaycastHit hit)
     {
         agent.destination = hit.point;
         Debug.Log("Lece!");
-        task = TaskList.Moving;
+   //     task = TaskList.Moving;
     }
 
-    public void Harvest(NavMeshAgent agent, RaycastHit hit, TaskList task, GameObject targetNode)
-    {
-        agent.destination = hit.collider.gameObject.transform.position;
-        Debug.Log("Wydobywam");
-        task = TaskList.Gathering;
-        targetNode = hit.collider.gameObject;
-    }
+   // public void Harvest(NavMeshAgent agent, RaycastHit hit, TaskList task, GameObject targetNode)
+   // {
+   //    agent.destination = hit.collider.gameObject.transform.position;
+   //    Debug.Log("Wydobywam");
+   //    task = TaskList.Gathering;
+   //    targetNode = hit.collider.gameObject;
+   // }
 }
