--- Game/Assets/Scripts/Unit.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cdbe3b2dbcdff294fa4822b9fbe399c4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
