+++ Game/Library/StateCache/SceneView/a119d0-mainStage.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":false,"pivot":{"x":-12.062980651855469,"y":18.169466018676759,"z":-22.745929718017579},"rotation":{"x":-0.3925285041332245,"y":0.04192414879798889,"z":-0.01789550669491291,"w":-0.9186376929283142},"size":8.30105209350586,"orthographic":false}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":false,"pivot":{"x":-12.856213569641114,"y":0.3135632276535034,"z":-6.519961357116699},"rotation":{"x":-0.3925285041332245,"y":0.04192414879798889,"z":-0.01789550669491291,"w":-0.9186376929283142},"size":13.358255386352539,"orthographic":false}
